+++ src/renderers/BaseRenderer.php
@@ -189,6 +189,11 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
     public $jsPositions = [View::POS_HEAD, View::POS_BEGIN, View::POS_END, View::POS_READY, View::POS_LOAD];
 
     /**
+     * @var bool add a new line to the beginning of the list, not to the end
+     */
+    public $prepend = false;
+
+    /**
      * @inheritdoc
      */
     public function setContext($context)
@@ -411,7 +416,8 @@ abstract class BaseRenderer extends BaseObject implements RendererInterface
             'max'               => $this->max,
             'min'               => $this->min,
             'attributes'        => $this->prepareJsAttributes(),
-            'indexPlaceholder'  => $this->getIndexPlaceholder()
+            'indexPlaceholder'  => $this->getIndexPlaceholder(),
+            'prepend'           => $this->prepend
         ], $this->jsExtraSettings));
 
         $js = "jQuery('#{$this->id}').multipleInput($options);";